Add about modal widget and test page
Test page doesn't quite work in PhantomJS, clicking the button from the
patternfly code example doesn't open the modal.

Local testing shows the widget methods and properties work as intended

Included title and trademark fields as widget properties

Couldn't find an xpath solution for key/value pairs in modal that are a
single <li> element with the key in <strong> block and value following.
Opted to use xpath to parse key, and then strip that key from the full
<li> text.