Fix problem with locator in SettingsNavDropdown
The code in CFME integration_tests triggers repr printing on SettingsNavDropdown,
which is a successor of NavDropdown, the successor of Widget.

Neither the Widget nor the NavDropdown is not defining the property `locator`,
yet the NavDropdown is trying to use it for repr, failing like this:

AttributeError: 'SettingsNavDropdown' object has no attribute 'locator'

This commit is setting this changing the repr to not use the 'locator'
but rather the id for stringification purposes.

Also some tests were added. The test_all_components contains code for
auto-discovering Widget classes, instantiating them in a programatically
defined view and then testing behavior of such objects.

Fixes: https://github.com/ManageIQ/integration_tests/issues/10191
This requires
https://github.com/RedHatQE/widgetastic.patternfly/pull/122 for tests to
fully finish passing.